move helper functions out
